move test helper module to const
