--- tests/components/nws/helpers/__init__.py
@@ -1,1 +1,0 @@
-"""Helpers for NWS tests."""
